пытается добавить отрисовку
