+++ js/main.js
@@ -1,3 +1,3 @@
 import './data.js';
 import './render.js';
-import './render-big-img.js';
+import './render-big-pic.js';
